Separate client tick to its own file
